writer: fix unsound `overwrite()` behavior
The data between `Writer/Reader` would go out-of-sync if this
function were used because the `Patch` that is supposed to
be applied on reclaimed `Reader` data in `push()` AFTER an
`overwrite()` was missing - since we `.clear()`'ed the `patches_old`
Vector.

Now, a `Patch` is inserted such that reclaimed data
will just be overwritten with a `.clone()`.